# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Arxivist Backend API
  version: 1.0.0
  description: |
    Бэкенд-модуль HeadHunter-агента (HeadHunter + интернет + базы знаний) с доступом по API-ключу.

    ## Аутентификация
    Все эндпоинты (кроме `/docs`, `/openapi.yaml`, `/health`) требуют API-ключ в заголовке
    `X-API-Key` (или `Authorization: Bearer <key>`). Ключи выпускаются оператором:
    `python scripts/apikeys.py create --label "<имя клиента>"`. Ключ показывается один раз;
    в хранилище лежит только его SHA-256. revoke действует немедленно, без рестарта.

    ## Два семейства эндпоинтов
    1. **Агентские задачи** (`/agent/jobs`) — полный LLM-агент со всеми интеграциями
       (HeadHunter, базы знаний, arXiv, веб-поиск, дашборды). Выполняются минуты, поэтому
       асинхронно: `POST` возвращает `202 {job_id}`, дальше поллинг `GET /agent/jobs/{id}`
       (рекомендуемый интервал 2–5 с). Результат — markdown-ответ агента.
    2. **Типизированные эндпоинты** (`/hh/*`, `/kb/*`, `/candidates/save`, `/arxiv/search`) —
       детерминированные вызовы без LLM, отвечают синхронно за секунды
       (исключение — `/arxiv/search` и `/kb/query`, до 1–3 минут).

    ## Коды ошибок
    Все не-2xx ответы имеют форму `{"error": {"code", "message", ...}}`:
    | HTTP | code | смысл |
    |---|---|---|
    | 400 | `invalid_request` | неверные параметры запроса |
    | 401 | `invalid_api_key` | ключ отсутствует/неверен/отозван |
...
    | 404 | `job_not_found` / `file_not_found` | нет такого объекта |
...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /files:
    post:
      tags: [files]
      summary: Загрузить документ (PDF/DOCX/MD/TXT, ≤25 МБ)
      description: Полученные `file_id` передаются в `attached_file_ids` агентской задачи.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: Метаданные сохранённых файлов
          content:
            application/json:
              example:
                files:
                  - { file_id: "0cd0424…", name: "vacancy.md", parse_mode: plain, char_count: 328, size: 508 }
                errors: []
        "401": { $ref: "#/components/responses/Unauthorized" }
  /files/{name}:
    get:
      tags: [files]
      summary: Скачать полученный PDF (например, статью arXiv)
      parameters:
        - { name: name, in: path, required: true, schema: { type: string }, example: "2401.12345v1.pdf" }
      responses:
        "200":
          description: PDF-файл
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /dashboards:
    get:
      tags: [meta]
      summary: Список сгенерированных дашбордов
      description: |
        HTML каждого дашборда доступен без ключа по секретному URL `/d/{uuid}`
        (относительно корня приложения, не /api/v1) — ссылку возвращает агент.
      responses:
        "200":
          description: Дашборды (новые первыми)
          content:
            application/json:
              example:
                dashboards:
                  - { uuid: "e3b0c44298fc1c149afbf4c8996fb924", title: "Обзор рынка Python", created_at: 1784200000, session_id: "api-…" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "Ключ вида axv_<id>_<secret>, выпущенный scripts/apikeys.py"
    BearerAuth:
      type: http
      scheme: bearer
      description: "Тот же ключ в Authorization: Bearer <key>"
  responses:
...
    Unauthorized:
      description: API-ключ отсутствует, неверен или отозван
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: invalid_api_key, message: "Provide a valid key in the X-API-Key header." } }
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
          additionalProperties: true
...